// src/api/generated/model/exerciseMuscleRoleInputRequestType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Fitness App API
 * Comprehensive API for fitness application with support for workout plans, exercises, and progress tracking. API provides OAuth2 authentication via Keycloak or JWT tokens.
 */

export type ExerciseMuscleRoleInputRequestType = typeof ExerciseMuscleRoleInputRequestType[keyof typeof ExerciseMuscleRoleInputRequestType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ExerciseMuscleRoleInputRequestType = {
  PRIMARY: 'PRIMARY',
  SECONDARY: 'SECONDARY',
} as const;